postgresql_saver/requirements.txt
numpy==2.0.2
openpyxl==3.1.5
pandas==2.2.3
requests==2.32.3
psycopg2-binary==2.9.11
confluent-kafka==2.10.0